Added a new Task to use pictures from picsum.photos to fill the Articles that don't have an image.

+++ Runners/modules/UserGeneratorRunner/Config/config.php
@@ -2,8 +2,8 @@
 
 return [
 
-    'max_new_users' => env('GENERATE_USERS_MAX', 20),
+    'max_new_users' => (int) env('GENERATE_USERS_MAX', 20),
 
-    'max_image_reuse' => env('GENERATE_USERS_MAX_IMAGE_REUSE', 50),
+    'max_image_reuse' => (int) env('GENERATE_USERS_MAX_IMAGE_REUSE', 50),
 
 ];
